requirements.txt
Flask
PyMuPDF
PyPDF2
pdf2image
reportlab
python-docx
Pillow
